Significantly Improve API + Optimize Decompression
This commit does a refactor of almost the entire API to break a lot of the confusing mishmash of results from a singular function into multiple functions. In addition, it makes decompression doable for buffers which are of equal size and allowed having a much cleaner and easy to use API there.